massive refactor on processor src/sink arch

+++ src/model/model.cpp
@@ -4,7 +4,7 @@
 
 #include "model.h"
 
-Model::Model(const ModelDesc &model_desc, Shape input_shape, int batch_size)
+Model::Model(const ModelDesc &model_desc, Shape input_shape, size_t batch_size)
     : model_desc_(model_desc),
       input_shape_(input_shape),
       batch_size_(batch_size) {}
